+++ modules/node-groups/main.tf
@@ -25,6 +25,7 @@ module "system_node_groups" {
   node_type                    = each.value.node_type
   node_type_version            = each.value.node_type_version
   node_size                    = each.value.node_size
+  ultra_ssd                    = each.value.ultra_ssd
   proximity_placement_group_id = each.value.proximity_placement_group_id
   fips                         = var.fips
   labels                       = merge(var.labels, each.value.labels)
@@ -65,6 +66,7 @@ module "user_node_groups" {
   node_type                    = each.value.node_type
   node_type_version            = each.value.node_type_version
   node_size                    = each.value.node_size
+  ultra_ssd                    = each.value.ultra_ssd
   proximity_placement_group_id = each.value.proximity_placement_group_id
   fips                         = var.fips
   labels                       = merge(var.labels, each.value.labels)